Modify models usage of the data api
Given the client for the data api (residing in the auth app) requires a
valid user token to be instantiated, we can no longer rely on it being
fully initialized on app boot.
boot.

This addresses the issue by removing the inheritance structure
descending from ApiModel, so the client can be initialized with a fresh
user token per request.

Since the client returns full fledged models, this extracts a dedicated
finder and a presenter for each model, in this case `organisation`.