Introduce Shifts resource
Shifts belong to a location and have an name, starting time and ending
time. The only required attributes so far are the location uid and
starting time.

This introduces a form object as well as two models dedicated to
converting start/end times into appropriate values. Since Rails
time_select helper introduces hidden date fields, we have to handle
separating those from the actual time of arrival. Another side effect of
the hidden fields is that is effectively not possible to have a "nil"
end time as require in this feature. The reason for that is that the
date will be passed to ActiveRecord which will convert the empty time
strings to a time of `0:00`. The solution here introduces two objects
responsible for coercing the params to values that allow the feature to
be implemented.