Extract ApiEnabledController
Controllers that required communication with the Auth API have started
to accumulate some similar methods.

This moves the methods related to instantiating the AuthApi client and
associated methods that make use of it to retrieve records, into a base controller.
Controllers that now require the client or any of the associated methods
can treated as extensions of ApiEnabledController.

In order to prevent regression, each of the specs for objects that
extend the extracted class now include a spec to verify the extension is
in place.