Refactor LocationShift
While `LocationShift` as a domain concept is relevant, the naming
convention for the underlying object was somewhat confusing. Looking at
the `LocationShift` model implementation, it delegates almost all
methods to `Shift` and decorates only three attributes (name,
starting_time and ending_time).

This makes the behaviour more explicit by having `LocationShift` extend
from `SimpleDelegator` and renaming it to `ShiftPresenter`. Even with
the tradeoff of having a pattern in the object name, this makes that
area of the system easier to understand.